Adding npm scripts for convenience
If `gulp` isn't installed globally on a system, the path to the local
module's "bin" directory was required.  Additionally, there was a slight
cognitive overhead with switching tools.  This keeps the focus on npm as
the single tool needed to be used for development. `npm run dev` will
now do the same thing as `gulp dev`, and if desired this could be
renamed to `npm start` since it seems like a development starting point.

Also added a `build` script set to `npm run build` which is equivalent
to running `gulp`.